requirements.txt
pillow==10.0.0
kafka-python==2.0.2
python-daemon
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.24.3
opencv-python==4.7.0.72
pytz==2023.3
sqlparse==0.4.4
django-ebhealthcheck==2.0.2
channels==4.0.0

